minor clean

+++ xcomm/src/atomic_rw/wrapper_type.hh
@@ -18,9 +18,9 @@ const u32 kInvalidSeq = 0;
 #pragma pack(1)
 template <typename T> struct alignas(64) WrappedType {
   // when the seq equals invalid seq, then it is being written
-  volatile u64 seq = kInvalidSeq + 1;
+  volatile u32 seq = kInvalidSeq + 1;
   T payload;
-  volatile u64 seq_check;
+  volatile u32 seq_check;
  public:
   WrappedType(const T &p) : payload(p), seq_check(seq) {}
 
@@ -44,17 +44,14 @@ template <typename T> struct alignas(64) WrappedType {
   inline void begin_write() {
     this->seq_check += 1;
     this->seq = kInvalidSeq;
-    //::r2::store_fence();
     ::r2::compile_fence();
   }
 
   inline void done_write() {
     // the update of *seqs* should after previous writes
-    //::r2::store_fence();
     ::r2::compile_fence();
     this->seq = this->seq_check;
     ::r2::compile_fence();
-    //r2::store_fence();
   }
 };
